// src/api/generated/model/draggableItem.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */

export interface DraggableItem {
  id?: string;
  text?: string;
  mediaUrl?: string;
  mediaType?: string;
  correctZones?: string[];
}