// src/api/generated/model/dropZone.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */

export interface DropZone {
  id?: string;
  label?: string;
  maxItems?: number;
  feedback?: string;
  position?: string;
}