// src/api/generated/model/dragAndDropOptions.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { DraggableItem } from "./draggableItem";
import type { DropZone } from "./dropZone";

export interface DragAndDropOptions {
  draggableItems?: DraggableItem[];
  dropZones?: DropZone[];
}